Add JWT authentication and role based access

+++ routes/post_routes.py
@@ -1,52 +1,66 @@
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy.orm import Session
-from app.database import SessionLocal
+
 from models.post_model import Post
 from schemas.post_schema import PostCreate
+from auth.dependencies import get_db, require_roles
+from models.user_model import User
 
 router = APIRouter(prefix="/posts", tags=["Posts"])
 
 
-def get_db():
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
-
-
-# Create post
 @router.post("/")
-def create_post(post: PostCreate, db: Session = Depends(get_db)):
+def create_post(
+    post: PostCreate,
+    db: Session = Depends(get_db),
+    current_user: User = Depends(require_roles("admin", "author"))
+):
     new_post = Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id
+        user_id=current_user.id
     )
+
     db.add(new_post)
     db.commit()
     db.refresh(new_post)
+
     return new_post
 
 
-# Get all posts
 @router.get("/")
 def get_posts(db: Session = Depends(get_db)):
-    posts = db.query(Post).all()
-    return posts
+    return db.query(Post).all()
 
 
-# Get post by id
 @router.get("/{post_id}")
 def get_post(post_id: int, db: Session = Depends(get_db)):
     post = db.query(Post).filter(Post.id == post_id).first()
+
+    if not post:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Post not found"
+        )
+
     return post
 
 
-# Delete post
 @router.delete("/{post_id}")
-def delete_post(post_id: int, db: Session = Depends(get_db)):
+def delete_post(
+    post_id: int,
+    db: Session = Depends(get_db),
+    current_user: User = Depends(require_roles("admin"))
+):
     post = db.query(Post).filter(Post.id == post_id).first()
+
+    if not post:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Post not found"
+        )
+
     db.delete(post)
     db.commit()
-    return {"message": "Post deleted"}
+
+    return {"message": "Post deleted successfully"}
